refactor(drug-class): sets drug in its own file for clarity
