+++ src/index.ts
@@ -1,4 +1,5 @@
-import { Drug, Pharmacy } from "./pharmacy";
+import { Pharmacy } from "./pharmacy";
+import { Drug } from './drug';
 import fs from "fs";
 
 const drugs = [
