+++ src/pharmacy.ts
@@ -1,14 +1,4 @@
-export class Drug {
-  name: string;
-  expiresIn: number;
-  benefit: number;
-
-  constructor(name: string, expiresIn: number, benefit: number) {
-    this.name = name;
-    this.expiresIn = expiresIn;
-    this.benefit = benefit;
-  }
-}
+import { Drug } from './drug';
 
 export class Pharmacy {
   drugs: Drug[];
@@ -19,18 +9,18 @@ export class Pharmacy {
   updateBenefitValue() {
     for (var i = 0; i < this.drugs.length; i++) {
       if (
-        this.drugs[i].name != "Herbal Tea" &&
-        this.drugs[i].name != "Fervex"
+        this.drugs[i].name != 'Herbal Tea' &&
+        this.drugs[i].name != 'Fervex'
       ) {
         if (this.drugs[i].benefit > 0) {
-          if (this.drugs[i].name != "Magic Pill") {
+          if (this.drugs[i].name != 'Magic Pill') {
             this.drugs[i].benefit = this.drugs[i].benefit - 1;
           }
         }
       } else {
         if (this.drugs[i].benefit < 50) {
           this.drugs[i].benefit = this.drugs[i].benefit + 1;
-          if (this.drugs[i].name == "Fervex") {
+          if (this.drugs[i].name == 'Fervex') {
             if (this.drugs[i].expiresIn < 11) {
               if (this.drugs[i].benefit < 50) {
                 this.drugs[i].benefit = this.drugs[i].benefit + 1;
@@ -44,14 +34,14 @@ export class Pharmacy {
           }
         }
       }
-      if (this.drugs[i].name != "Magic Pill") {
+      if (this.drugs[i].name != 'Magic Pill') {
         this.drugs[i].expiresIn = this.drugs[i].expiresIn - 1;
       }
       if (this.drugs[i].expiresIn < 0) {
-        if (this.drugs[i].name != "Herbal Tea") {
-          if (this.drugs[i].name != "Fervex") {
+        if (this.drugs[i].name != 'Herbal Tea') {
+          if (this.drugs[i].name != 'Fervex') {
             if (this.drugs[i].benefit > 0) {
-              if (this.drugs[i].name != "Magic Pill") {
+              if (this.drugs[i].name != 'Magic Pill') {
                 this.drugs[i].benefit = this.drugs[i].benefit - 1;
               }
             }
